+++ packages/adminui-vue/components/UTableEntity/components/UTableEntityRoot.vue
@@ -464,7 +464,7 @@
               <u-dropdown-item
                 icon="u-icon-file-preview"
                 label="ChangesHistory"
-                @click="showRevision(selectedRowId)"
+                @click="showRecordHistory(selectedRowId)"
               />
             </template>
           </slot>
@@ -662,7 +662,7 @@ export default {
       'createLink',
       'audit',
       'createNewVersion',
-      'showRevision'
+      'showRecordHistory'
     ]),
     ...mapMutations(['SELECT_COLUMN', 'SELECT_ROW']),
     getCellTemplate (column) {