+++ packages/adminui-vue/components/UTableEntity/store.js
@@ -2,6 +2,7 @@ const UB = require('@unitybase/ub-pub')
 const { Notification: $notify } = require('element-ui')
 const uDialogs = require('../../utils/uDialogs')
 const { exportExcel, exportCsv, exportHtml } = require('../../utils/fileExporter')
+const helpers = require('../../utils/Form/helpers')
 const lookups = require('../../utils/lookups')
 const AUDIT_ENTITY = 'uba_auditTrail'
 const Vue = require('vue')
@@ -393,9 +394,17 @@ module.exports = (instance) => ({
       }
     },
 
-    async refresh ({ commit, dispatch }) {
-      commit('PAGE_INDEX', 0)
-      await dispatch('fetchItems')
+    async refresh ({ commit, dispatch, getters }) {
+      commit('LOADING', true)
+      try {
+        commit('PAGE_INDEX', 0)
+        for (const { entity, associatedAttr } of getters.lookupEntities) {
+          await lookups.refresh(entity, [associatedAttr])
+        }
+        await dispatch('fetchItems')
+      } finally {
+        commit('LOADING', false)
+      }
     },
 
     async getTotal ({ commit, dispatch }) {
@@ -741,21 +750,8 @@ module.exports = (instance) => ({
       }
     },
 
-    async showRevision ({ getters }, ID) {
-      const { mi_data_id: historyId } = await UB.Repository(getters.entityName)
-        .attrs('ID', 'mi_data_id')
-        .selectById(ID)
-
-      UB.core.UBApp.doCommand({
-        cmdType: 'showList',
-        cmdData: {
-          entityName: getters.entityName,
-          columns: getters.columns.concat('mi_dateTo', 'mi_dateFrom')
-        },
-        isModal: instance.isModal,
-        instanceID: historyId,
-        __mip_recordhistory: true
-      })
+    async showRecordHistory ({ getters }, ID) {
+      return helpers.showRecordHistory(getters.entityName, ID, getters.repository().fieldList, getters.columns)
     }
   }
 })