+++ packages/ubq/CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 ### Changed
+ - `ubq_messages` nav shortcut - explicitly define fieldList as `["mi_createDate", "queueCode", "completeDate", "msgCmd"]` 
+   for better UX
+ - **mail job** will retrieve ubq_messages.msgData CLOB (a message body) one record at once.
+   This prevents `<<<ERR_RESOURCE_LIMITS_EXCEED>>>` error in case overall size of mail bodies for 100 queued messages
+   exceed 50Mb [UBDF-14156] 
 
 ### Deprecated